Changed pressure unit conversion from inches of mercury to millibar, added unit tests for conversion.
    Converting millipascals to inches mercury proved difficult to handle the precision, millibar is a much cleaner conversion and easily testable.

+++ types/types.go
@@ -17,7 +17,7 @@ type TippingBucket struct {
 type Bme280Readings struct {
 	TempF    float32 `json:"tempF"`
 	Humidity float32 `json:"humidity"`
-	Pressure float32 `json:"pressure"`
+	Pressure float64 `json:"pressure"`
 }
 
 const (
